Revert the merge of two pull-requests that target CGAL-4.14-branch
- Revert "Merge pull request #3857 from MaelRL/CGAL-Fix_duplicate_non_manifold_vertices-GF"
- Revert "Merge pull request #3898 from lrineau/Mesh_3-stop_during_protection-lrineau"

+++ Kernel_d/include/CGAL/Kernel_d/DirectionHd_impl.h
@@ -52,8 +52,8 @@ cmp(const DirectionHd<RT,LA>& h1,
   int c2 = CGAL_NTS sign(h2.delta(i)); 
   if (c1 != c2) return CGAL_NTS compare(c1,c2); 
  
-  RT s1 = (RT) (int) CGAL_NTS sign(h2.delta(i)) * h2.delta(i);
-  RT s2 = (RT) (int) CGAL_NTS sign(h1.delta(i)) * h1.delta(i);
+  RT s1 = (RT) CGAL_NTS sign(h2.delta(i)) * h2.delta(i); 
+  RT s2 = (RT) CGAL_NTS sign(h1.delta(i)) * h1.delta(i); 
 
   i++; 
   Comparison_result c;
